[Matrix] Multiplying instead of adding
The way to compute the global matrix was adding the submatrices. The unknown
values used to be maximized to avoid false positive. But it was to much
penalizing. Now, for unknown values, we set 1, then all matrices are multiplied.
Thus, if a value in unknown, it's not penalizing whereas if the value is known,
it's positive.