[minhashing] consume less memory
Storing the whole document matrix was useless because it was a boolean and
sparse matrix. So the only stored element are equal to one, and therefore it's
useless to store the data list : only the position are interesting.

During the signature step, the read lines are useless because they aren't use
anymore, so they are deleted to save memory.