[aligner] Normalize the sets before calling the findneigbours() function
The treaments must be applied before the findneighbours() function is
called, otherwise it's a little bit useless. But, the treatments are
also applied by align() function. It's too much and useless.
We should find a way to apply it once whatever is the called function
(because align() can be called lonely?)

The *temporary* solution is to call normalize_set twice, but it has to
be just a *temporary* solution !