[dataio] Add a ``rqlquery()`` function
Execute a rql query on a given cubicweb host, and parse and format the results
to the data structure used by nazca alignement process.

The cubicweb host is expected to return a csv file (thanks to the csvexport
view)

***

[dataio] ``rqlquery()`` handles indexes formatting.

The output stream get by urllib in ``rqlquery`` is forwarded to the
``parsefile`` function, which already support indexes formatting.